requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-cors==3.0.10
python-dotenv==0.19.0
pandas==1.3.3
numpy==1.21.2
transformers==4.11.3
torch==1.9.0
powerbiclient==0.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5 